pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-aqi-analytics-pipeline"
version = "0.1.0"
description = "Weather and Air Quality Analytics Pipeline with dbt and Docker"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core dependencies
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.7",
    "alembic>=1.11.0",
    
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Scheduling (optional)
    "schedule>=1.2.0",
    
    # Development and testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.14.0",
    "pre-commit>=3.3.0",
]

airflow = [
    "apache-airflow>=2.7.0",
    "apache-airflow-providers-postgres>=5.6.0",
]

prefect = [
    "prefect>=2.11.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/weather-aqi-analytics-pipeline"
Repository = "https://github.com/yourusername/weather-aqi-analytics-pipeline.git"
Issues = "https://github.com/yourusername/weather-aqi-analytics-pipeline/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | dbt_weather/target
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "dist",
    "build",
    "dbt_weather/target",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "dbt_weather/",
    "tests/",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src"
testpaths = [
    "tests",
]